feat: final version of investor edit profile form
