+++ src/domain/profile/atoms/investor-edit-profile-form/investor-edit-profile-form.defaults.ts
@@ -1,5 +1,3 @@
-import { IEditProfileFormFieldsData } from '../../profile.types';
-
 export const translationStrings = [
   'common.form.field.error.required',
   'common.buttons.reset',
@@ -7,7 +5,6 @@ export const translationStrings = [
   'common.form.field.error.contain.single.word',
   'common.form.field.error.contain.enters.or.spaces',
   'common.form.field.error.starts.or.ends.with.whitespace',
-  'common.errors.something.went.wrong',
   'common.focus.markets.b2b',
   'common.focus.markets.b2c',
   'common.industrial.sectors.web3',
@@ -86,23 +83,4 @@ export const translationStrings = [
   'component.dashboard.edit.profile.form.demand.field.label',
   'component.dashboard.edit.profile.form.why.startup.should.match.with.you.field.label',
   'component.dashboard.edit.profile.form.why.startup.should.match.with.you.field.max.length.error',
-  'component.dashboard.edit.profile.form.messages.success',
 ] as const;
-
-export const defaultFormFieldsValues: IEditProfileFormFieldsData = {
-  companyName: '',
-  contactEmail: '',
-  firstName: '',
-  lastName: '',
-  location: '',
-  images: [],
-  focusMarketIds: [],
-  industrialSectorIds: [],
-  investmentSizeIds: [],
-  investmentStageTypeIds: [],
-  investorDemandTypeIds: [],
-  investorProfileTypeId: '',
-  startupSectorIds: [],
-  teamSizeIds: [],
-  whyStartupShouldMatchWithYou: '',
-};
