+++ src/domain/profile/atoms/investor-edit-profile-form/investor-edit-profile-form.styles.ts
@@ -1,7 +1,13 @@
 import styled from '@emotion/styled';
 
+import { respondToMax } from '@infrastructure';
+
 const StyledWrapper = styled.form`
-  max-width: 1400px;
+  max-width: 850px;
+
+  ${respondToMax.xmobile`
+    max-width: 100%;
+  `}
 `;
 
 const S = {
