+++ src/domain/profile/atoms/investor-edit-profile-form/investor-edit-profile-form.types.ts
@@ -1,3 +1,6 @@
+import { FormHTMLAttributes } from 'react';
+import { Control } from 'react-hook-form';
+
 import {
   IFocusMarket,
   IIndustrialSector,
@@ -5,19 +8,23 @@ import {
   IInvestmentStageType,
   IInvestorDemandType,
   IInvestorProfileType,
-  IProfileDetails,
   IStartupSector,
   ITeamSize,
 } from '@interfaces';
 
-export interface IDesktopInvestorEditProfileFormProps {
+import { IEditProfileFormFieldsData } from '../../profile.types';
+
+export interface IInvestorEditProfileFormProps extends FormHTMLAttributes<HTMLFormElement> {
+  control: Control<IEditProfileFormFieldsData>;
   focusMarkets: IFocusMarket[];
   industrialSectors: IIndustrialSector[];
   investmentSizes: IInvestmentSize[];
   investmentStageTypes: IInvestmentStageType[];
   investorDemandTypes: IInvestorDemandType[];
   investorProfileTypes: IInvestorProfileType[];
-  profileDetails: IProfileDetails | undefined;
+  isDirty: boolean;
+  isLoading: boolean;
+  onResetButtonClick: () => void;
   startupSectors: IStartupSector[];
   teamSizes: ITeamSize[];
 }
