+++ src/domain/profile/molecules/subscription-plans/subscription-plans.styles.ts
@@ -12,6 +12,7 @@ const StyledHeader = styled(Typography)`
   `}
 
   ${({ theme }) => respondToMax.xmobile`
+    text-align: center;
     margin-bottom: ${theme.spacing(1)};
   `}
 `;
